app/requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
lxml==4.9.2
schedule==1.1.0
twilio
python-dotenv
